refactor(registration-service): removed invalidation flag, added methods for consents federation, new pagination logic for multiple repositories, moved transactions to data layer

+++ common/ASC.Identity/registration/registration-core/registration-service/src/main/java/com/asc/registration/service/transfer/request/create/CreateTenantClientCommand.java
@@ -1,4 +1,4 @@
-// (c) Copyright Ascensio System SIA 2009-2024
+// (c) Copyright Ascensio System SIA 2009-2025
 //
 // This program is a free software product.
 // You can redistribute it and/or modify it under the terms
@@ -53,7 +53,7 @@ public class CreateTenantClientCommand implements Serializable {
   /** The ID of the tenant. Must be greater than or equal to 1. */
   @Min(value = 1, message = "tenant id must be greater than or equal to 1")
   @JsonProperty("tenant_id")
-  private int tenantId;
+  private long tenantId;
 
   /** The name of the client. Must not be empty and should be between 3 and 256 characters. */
   @NotEmpty(message = "client name must not be empty")
